More descriptive file names in frontend
