+++ src/tasks/start/frontend/watch-app.ts
@@ -1,0 +1,8 @@
+import execa from 'execa'
+
+/**
+ * Calls the app's front end build watcher.
+ */
+export async function startAppWatcher(appSrcPath: string): Promise<void> {
+  await execa('npm', ['run', 'watch'], { cwd: appSrcPath })
+}
