--- test/src/tasks/start/frontend/app.test.ts
@@ -1,5 +1,0 @@
-import { assert } from 'chai'
-
-describe('app.ts', function() {
-  it.skip('more tests needed', async function() {})
-})
