--- test/src/tasks/start/frontend/client.test.ts
@@ -1,5 +1,0 @@
-import { assert } from 'chai'
-
-describe('client.ts', function() {
-  it.skip('more tests needed', async function() {})
-})
