--- test/src/tasks/start/frontend/frontend.test.ts
@@ -1,5 +1,0 @@
-import { assert } from 'chai'
-
-describe('frontend.ts', function() {
-  it.skip('more tests needed', async function() {})
-})
